Rely on community.docker collection (#58)
* Assure ansible has docker_container module installed (#26)

We need ansible with docker module for using this module.

* Bulk fixes

Building all these containers does not really depend on python
version of ansible version. This moves this test to a standalone pipeline.

Starting with this version we only support and require latest community.docker collection and we assume user installed it correctly.

Installing just ansible distribution is not enough to make this
plugin functional.